find or create resource config and cache without descriptor
Removed the resource config and cache descriptors! We were able to do
that because we are now passing in the resource cache for the custom
type (if it uses one) to the FindOrCreateResourceConfig or
FindOrCreateResourceCache method, and with the resource cache we no
longer require the descriptor to tell us what its chain of parent types
are. By doing this, we can remove versioned resource types from the step
plans (such as get, check and put), we need to remove its usage from the
steps itself.

Currently FindCheckContainer is broken because it also used
FindOrCreateResourceConfig, but now that it takes in a resource cache we
dont have a resource cache to pass into it from the FindCheckContainer
method. We will be removing the need to get a resource config at all in
the future for FindCheckContainer so we can keep it broken for now.

Lastly, since we no longer have a ResourceCacheDescriptor, we don't need
the name of the resource cache object to be UsedResourceCache. (yay)

[#6274]

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>