B: remove unnecessary error
technically this is a behaviour change since it modifies the error
message, but that's the extent of it.
`ErrResourceConfigCheckSessionExpired` also has a similar (and arguably
less helpful) error message than `db.ContainerOwnerDisappearedError`,
and doesn't even mention anything about resource config check sessions -
just a generic error for all owner types.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>