B: backport caching streamed volumes
this commit backports #6660, which creates a worker_resource_cache on
the worker that a resource volume was streamed into. it does this by:

1. streaming the volume into an empty volume (like before). this volume
   becomes the resource cache (if the src volume was also a resource
   cache)
2. creating a COW volume over that empty volume and mounting that. this
   is to avoid mutating the cached volume.

that PR also introduces a change to the get step - if there is a cached
volume on any worker, use that, rather than re-running the get step.

some modifications were made over the original implementation:

* for "get from local cache", previously the get step would lookup all
  of the workers that contain the resource cache, and then walk through
  the list of workers and try to grab the volume from each one,
  stopping at the first success. now, more of this logic is pushed into
  the worker pool - the get step simply asked the pool for a resource
  cache volume (rather than the list of workers)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>