check factory resource types grouped by pipelineID
The ResourceTypes method on the check factory will now return back a map
of pipelineID to resource types in order to easily pass in the correct
list of resource types for a specific pipeline. This fixes a bug in the
CheckPlan method, where it expects the list of resource types to be
already condensed to a list of resource types within the specific
pipeline.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>
Co-authored-by: Rui Yang <ryang@pivotal.io>